first commit #32

+++ Library/DataStructure/UnionFind.hpp
@@ -3,9 +3,9 @@
 /**
  * @file UnionFind.hpp
  * @author log K (lX57)
- * @brief UnionFind - 素集合データ構造
- * @version 4.0
- * @date 2024-09-05
+ * @brief Union-Find - 素集合データ構造
+ * @version 5.0
+ * @date 2025-02-07
  */
 
 #include "../Common.hpp"
@@ -13,15 +13,16 @@
 class UnionFind{
     public:
     /**
-     * @brief 要素数 `size` でUnionFindを初期化する。
+     * @brief 要素数 `size` で Union-Find を初期化します。
+     * @note 要素番号は 0-index です。
      * @param size 要素数
      */
-    UnionFind(int size) : data_(size, -1){}
+    UnionFind(size_t size) : data_(size, -1){}
 
     /**
-     * @brief 要素 `k` の親を返す。
+     * @brief 要素 `k` が属する集合の代表元を返します。
      * @param k 探索する要素
-     * @return int 親要素の番号
+     * @return int 要素 `k` が属する集合の代表元
      */
     int Find(int k){
         if(data_[k] < 0) return k;
@@ -30,18 +31,18 @@ class UnionFind{
     }
 
     /**
-     * @brief 要素 `x` と要素 `y` が同じ集合に属しているかを判定する。
-     * @param x 要素番号 (0-index)
-     * @param y 要素番号 (0-index)
+     * @brief 要素 `x` と要素 `y` が同じ集合に属しているかを判定します。
+     * @param x 要素番号
+     * @param y 要素番号
      */
     bool Same(int x, int y){
         return Find(x) == Find(y);
     }
 
     /**
-     * @brief 要素 `x` と要素 `y` を併合する。
-     * @param x 要素番号 (0-index)
-     * @param y 要素番号 (0-index)
+     * @brief 要素 `x` と要素 `y` を併合します。併合した場合は `true` を返します。
+     * @param x 要素番号
+     * @param y 要素番号
      * @return true 要素 `x` と要素 `y` を併合した (元々未併合だった)
      * @return false 要素 `x` と要素 `y` が既に併合済だった
      */
@@ -55,7 +56,7 @@ class UnionFind{
     }
 
     /**
-     * @brief 各集合に所属する要素を列挙する。
+     * @brief 各集合に所属する要素を列挙します。
      * @return vector<vector<int>> 各集合に所属する要素の一覧
      */
     vector<vector<int>> Group(){
@@ -70,8 +71,8 @@ class UnionFind{
     }
 
     /**
-     * @brief 要素 `k` が属する集合の要素数を求める。
-     * @param k 要素番号 (0-index)
+     * @brief 要素 `k` が属する集合の要素数を求めます。
+     * @param k 要素番号
      * @return int 集合の要素数
      */
     int Size(int k){
@@ -80,5 +81,5 @@ class UnionFind{
     }
 
     private:
-    vector<int> data_;
+    vector<int> data_; // 代表元は-(集合の要素数)を、非代表元は代表元の番号を経路圧縮により持つ
 };
